src/backend/app/requirements.txt
adagio==0.2.4
aiofiles==23.1.0
aiosqlite==0.19.0
alembic==1.13.0
altair==5.2.0
ansi2html==1.8.0
antlr4-python3-runtime==4.11.1
anyio==3.7.0
app==0.0.1
appdirs==1.4.4
asttokens==2.2.1
asyncpg==0.28.0
attrs==23.1.0
axios==0.4.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
boto3==1.33.6
botocore==1.33.6
catboost==1.2.2
category-encoders==2.6.3
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.8.0
colour==0.1.5
comm==0.1.4
contourpy==1.1.0
cryptography==41.0.7
cycler==0.11.0
Cython==3.0.6
daal==2024.0.1
daal4py==2024.0.1
dacite==1.8.1
dash==2.14.2
dash-auth==2.0.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-testing-stub==0.0.2
dask==2023.7.0
databases==0.6.2
databricks-cli==0.18.0
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
distributed==2023.7.0
dnspython==2.4.2
docker==6.1.3
docutils==0.20.1
dotty-dict==1.3.1
dtreeviz==2.2.2
email-validator==2.0.0.post2
entrypoints==0.4
evidently==0.2.8
executing==1.2.0
explainerdashboard==0.4.3
fairlearn==0.7.0
fastapi==0.89.1
fastjsonschema==2.19.0
ffmpy==0.3.1
filelock==3.12.2
Flask==2.2.3
flask-simplelogin==0.1.2
Flask-WTF==1.2.1
fonttools==4.40.0
fs==2.4.16
fsspec==2023.12.0
fugue==0.8.6
fugue-sql-antlr==0.1.8
future==0.18.3
gevent==23.9.1
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.2
gradio==3.50.2
gradio_client==0.6.1
graphviz==0.20.1
greenlet==3.0.1
h11==0.14.0
html5tagger==1.3.0
htmlmin==0.1.12
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.19.4
hyperopt==0.2.7
idna==3.4
ImageHash==4.3.1
imageio==2.33.0
imbalanced-learn==0.11.0
importlib-metadata==5.2.0
importlib-resources==6.1.1
iniconfig==2.0.0
interpret==0.4.4
interpret-core==0.4.4
invoke==1.7.3
ipykernel==6.25.1
ipython==8.14.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-dash==0.4.2
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.9
kaleido==0.2.1
keyring==24.0.0
kiwisolver==1.4.4
kmodes==0.12.2
lazy_loader==0.3
lightgbm==4.1.0
lime==0.2.0.1
llvmlite==0.41.1
locket==1.0.0
lxml==4.9.3
m2cgen==0.10.0
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.6.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mlflow==1.30.1
mlxtend==0.23.0
more-itertools==9.1.0
moto==4.2.11
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multimethod==1.10
multiprocess==0.70.15
mysql==0.0.3
mysqlclient==2.2.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.7.0.72
optuna==3.4.0
orjson==3.9.10
ormar==0.12.1
outcome==1.3.0.post0
oyaml==1.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
partd==1.4.1
patsy==0.5.4
percy==2.0.2
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.10.0
plotly==5.18.0
plotly-resampler==0.9.1
pluggy==1.3.0
pmdarima==2.0.4
postgrest==0.10.6
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.5
psycopg2-binary==2.9.5
pure-eval==0.2.2
py4j==0.10.9.7
pyaml==23.9.7
pyarrow==13.0.0
pycaret==3.2.0
pycparser==2.21
pydantic==1.10.4
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.8.0
PyMySQL==1.0.3
pynndescent==0.5.11
pyod==1.1.2
pyodbc==5.0.1
pyOpenSSL==23.3.0
pyparsing==3.1.0
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-gitlab==3.15.0
python-multipart==0.0.6
python-semantic-release==7.33.2
pytz==2022.7.1
PyWavelets==1.5.0
pywin32==306
pywin32-ctypes==0.2.1
PyYAML==6.0
pyzmq==25.1.1
qpd==0.4.4
querystring-parser==1.2.4
readme-renderer==40.0
realtime==1.0.0
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
responses==0.24.1
retrying==1.3.4
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.13.2
s3transfer==0.8.2
SALib==1.4.7
sanic==23.3.0
sanic-routing==22.8.0
schemdraw==0.15
scikit-base==0.5.2
scikit-image==0.22.0
scikit-learn==1.2.2
scikit-learn-intelex==2024.0.1
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
selenium==4.2.0
semantic-version==2.10.0
semver==2.13.0
shap==0.43.0
six==1.16.0
skope-rules==1.0.1
sktime==0.21.1
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.41
sqlglot==19.8.2
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.22.0
statsforecast==1.5.0
statsmodels==0.14.0
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supafunc==0.2.2
sympy==1.12
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tbb==2021.11.0
tblib==3.0.0
tenacity==8.2.3
threadpoolctl==3.2.0
tifffile==2023.9.26
tomlkit==0.11.8
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.65.0
trace-updater==0.0.9.1
tracerite==1.1.0
traitlets==5.9.0
treeinterpreter==0.2.3
triad==0.9.1
trio==0.23.1
trio-websocket==0.11.1
tsdownsample==0.1.2
twine==3.8.0
typeguard==4.1.5
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.120
umap-learn==0.5.5
urllib3==1.26.18
urllib3-secure-extra==0.1.0
uvicorn==0.20.0
visions==0.7.5
waitress==2.1.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.9
wordcloud==1.9.2
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.1
xgboost==1.7.6
xmltodict==0.13.0
xxhash==3.4.1
ydata-profiling==4.6.0
yellowbrick==1.5
zict==3.0.0
zipp==3.15.0
zope.event==5.0
zope.interface==6.1
